fix: prevent creating duplicate machine ids
We saw the same machine id created more than one time.
This caused an issue where the cli would not be able
to connect to the correct machine.

We now wait for the instance to be created.
We don't care about what state the instance is in
only that it has been handled by AWS.

We believe duplication would happen when the scheduleTask
would finish quickly, receive the next update from the API,
but without AWS reporting the machine as pending.

Co-authored-by: Billy Batista <bootlegbilly@protonmail.ch>
Signed-off-by: Chris Goller <goller@gmail.com>